Refactor window classes: Remove setClosable() method calls
The setClosable() method calls were removed from the OrganizerWindow and ConferenceWindow classes. This method was unnecessary and caused issues with the window functionality. The commit refactors the code to remove the setClosable() method calls and improves the overall code quality.
Style Changes in LogInPanel (size changes, system pop ups, and pwd field).
PopUpWindow was created to handle pop up alerts, with posibility of success, error, and warning.
Bug fixes in NonEditableModel.